renamed resources to wiki
