+++ Portal/src/Datahub.Portal/Components/EmbeddedWiki.razor
@@ -1,5 +1,4 @@
-﻿@using Datahub.Core.Services.Resources
-@using Datahub.Core.Services.Wiki;
+﻿@using Datahub.Core.Services.Wiki;
 @inject IWikiService _resourcesService
 @inject CultureService _cultureService
 @inject NavigationManager _navManager
