+++ Portal/src/Datahub.Portal/Startup.cs
@@ -59,7 +59,6 @@ using Microsoft.Extensions.Azure;
 using Microsoft.Identity.Web.UI;
 using Tewr.Blazor.FileReader;
 using Datahub.Core.Services.ResourceManager;
-using Datahub.Core.Services.Resources;
 
 [assembly: InternalsVisibleTo("Datahub.Tests")]
